+++ src/__init__.py
@@ -1,6 +1,6 @@
 """Deep Research From Scratch - Tutorial implementation."""
 
-from .logging import setup_logging as _setup_logging
+from deep_research.logging import setup_logging as _setup_logging
 
 # Configure logging on package import (idempotent)
 _setup_logging()
