+++ src/deep_research/research_agent_scope.py
