+++ src/deep_research/state_scope.py
